fix(benchmark): decouple analyze from aea-heavy tournament import
PR #294 added `from benchmark.tournament import TOURNAMENT_TOOLS_JSON,
load_tournament_tools` to analyze.py. tournament.py imports ipfs_loader,
dotenv, and KeyChain at module level, so this pulled the entire
open-autonomy / aea stack into analyze's import graph.

The benchmark_flywheel job runs analyze in a minimal env
(`pip install requests scipy numpy`), so the import now crashes at
`import yaml` (and then `aea`) before doing any work, failing every
`python -m benchmark.analyze` invocation.

Move the two symbols analyze actually needs (TOURNAMENT_TOOLS_JSON, a
Path constant, and load_tournament_tools, a stdlib JSON reader) into a
new dependency-free module benchmark/tournament_tools.py. analyze and
the config test import from there; tournament.py re-imports
load_tournament_tools for its internal use. The benchmark job's
lightweight install stays valid.
